Adjust UI of Breakpoints
Fixes #3825. Fixes #3340.

The thinking here is that we should have an absolute minimum of
difference between loading, error, warning, and normal states.

- Break the various PanelSummary components out into their own files.
  This component is surprisingly complex, and has been a bit of a
  hot-point for changes. I've tried to give the logic that is there a
  little more room to breathe by pulling the components rendered out
  into nearby files, to remove their behavior from the cognitive load
  completely when thinking about how breakpoints behave.
- The comment button is now always visible, as is the timeline, and
  `leading-none` has been added to the Material Icon that is shown on
  warning, so that it doesn't increase the size a couple of pixels when
  it gets rendered (this creates a small jump and makes the heights
  inconsistent).
- Disabled navigation buttons will no longer turn the cursor into a pointer

https://user-images.githubusercontent.com/5903784/136997361-37ce0cdd-c397-4fac-a32b-997c94500240.mp4
